Simplified small range code, and added fast path
The small range codepath was complicated, slow, and unnecessary for small-width types. It has been streamlined to run only when necessary, and to do less work when it does run. In particular, “small range” now means, essentially, “the same or adjacent raw binades, close enough not to overflow `significandBitCount` when measured in terms of the smallest `ulp` in the range”.

The multiple-binade codepaths that used spare bits have been eliminated because they were not pulling their weight.

Also, a new fast path has been added to handle ranges that contain only full binades and are either bounded by or symmetric about zero.